pyproject.toml
[tool.poetry]
name = "level-3-text2sql-chatbot"
version = "0.1.0"
description = ""
authors = ["Arman Tunga <armantunga@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
haystack-ai = "^2.4.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
